Add mpl- prefix and rename packages
Also add descriptions to JS packages.
