+++ README.md
@@ -1,5 +1,7 @@
-# Token Authorization Rules
+# Metaplex Token Authorization Rules
 A program that provides the ability to create and execute rules to restrict common token operations such as transferring and selling.
+
+> ⚠️ **Metaplex Token Authorization Rules is currently experimental and has not been formally audited. Use in production at your own risk.**
 
 ## Overview
 Authorization rules are variants of a `Rule` enum that implements a `validate()` function.
@@ -20,7 +22,7 @@ use solana_sdk::{
     commitment_config::CommitmentConfig, native_token::LAMPORTS_PER_SOL, signature::Signer,
     signer::keypair::Keypair, transaction::Transaction,
 };
-use token_authorization_rules::{
+use mpl_token_auth_rules::{
     state::{Operation, Rule, RuleSet},
     Payload, PayloadVec,
 };
@@ -46,7 +48,7 @@ fn main() {
     println!("Payer balance: {}", balance);
 
     // Find RuleSet PDA.
-    let (ruleset_addr, _ruleset_bump) = token_authorization_rules::pda::find_ruleset_address(
+    let (ruleset_addr, _ruleset_bump) = mpl_token_auth_rules::pda::find_ruleset_address(
         payer.pubkey(),
         "da rulez".to_string(),
     );
@@ -81,8 +83,8 @@ fn main() {
         .unwrap();
 
     // Create a `create` instruction.
-    let create_ix = token_authorization_rules::instruction::create(
-        token_authorization_rules::id(),
+    let create_ix = mpl_token_auth_rules::instruction::create(
+        mpl_token_auth_rules::id(),
         payer.pubkey(),
         ruleset_addr,
         "da rulez".to_string(),
@@ -107,8 +109,8 @@ fn main() {
     let payload = Payload::new(None, None, Some(2), None);
 
     // Create a `validate` instruction.
-    let validate_ix = token_authorization_rules::instruction::validate(
-        token_authorization_rules::id(),
+    let validate_ix = mpl_token_auth_rules::instruction::validate(
+        mpl_token_auth_rules::id(),
         payer.pubkey(),
         ruleset_addr,
         "da rulez".to_string(),
@@ -146,15 +148,19 @@ fn main() {
 
 ---
 
-### Build the rust program alone
+### Build and test the Rust program
 ```
-$ yarn build:rust
+$ cd program/
+$ cargo build-bpf
+$ cargo test-bpf
+$ cd ..
 ```
 
 ---
 
-### Generate the JS SDK and rebuild IDL only (using shank and solita)
+### Build the program, generate the JS API, and rebuild IDL (using Shank and Solita)
 ```
+$ yarn build:rust
 $ yarn solita
 ```
