+++ cli/src/helpers/pda.ts
@@ -1,5 +1,5 @@
 import { PublicKey } from "@solana/web3.js";
-import { PREFIX, PROGRAM_ID } from "../../../packages/sdk/src/token-authorization-rules";
+import { PREFIX, PROGRAM_ID } from "../../../packages/sdk/src/mpl-token-auth-rules";
 
 export const findRuleSetPDA = async (payer: PublicKey, name: string) => {
     return await PublicKey.findProgramAddress(
