+++ packages/sdk/src/mpl-token-authorization-rules.ts
@@ -1,0 +1,5 @@
+export * from './errors';
+// @ts-ignore
+export * from './generated';
+
+export const PREFIX = 'ruleset';
