+++ program/tests/integration.rs
@@ -2,14 +2,14 @@
 
 pub mod utils;
 
+use mpl_token_auth_rules::{
+    state::{Operation, Rule, RuleSet},
+    Payload,
+};
 use rmp_serde::Serializer;
 use serde::Serialize;
 use solana_program_test::tokio;
 use solana_sdk::{signature::Signer, transaction::Transaction};
-use token_authorization_rules::{
-    state::{Operation, Rule, RuleSet},
-    Payload,
-};
 use utils::program_test;
 
 #[tokio::test]
@@ -17,7 +17,7 @@ async fn test_validator_transaction() {
     let mut context = program_test().start_with_context().await;
 
     // Find RuleSet PDA.
-    let (ruleset_addr, _ruleset_bump) = token_authorization_rules::pda::find_ruleset_address(
+    let (ruleset_addr, _ruleset_bump) = mpl_token_auth_rules::pda::find_ruleset_address(
         context.payer.pubkey(),
         "da rulez".to_string(),
     );
@@ -52,8 +52,8 @@ async fn test_validator_transaction() {
         .unwrap();
 
     // Create a `create` instruction.
-    let create_ix = token_authorization_rules::instruction::create(
-        token_authorization_rules::id(),
+    let create_ix = mpl_token_auth_rules::instruction::create(
+        mpl_token_auth_rules::id(),
         context.payer.pubkey(),
         ruleset_addr,
         "da rulez".to_string(),
@@ -79,8 +79,8 @@ async fn test_validator_transaction() {
     let payload = Payload::new(None, None, Some(2), None);
 
     // Create a `validate` instruction.
-    let validate_ix = token_authorization_rules::instruction::validate(
-        token_authorization_rules::id(),
+    let validate_ix = mpl_token_auth_rules::instruction::validate(
+        mpl_token_auth_rules::id(),
         context.payer.pubkey(),
         ruleset_addr,
         "da rulez".to_string(),
