+++ program/tests/utils/mod.rs
@@ -1,9 +1,5 @@
 use solana_program_test::ProgramTest;
 
 pub fn program_test() -> ProgramTest {
-    ProgramTest::new(
-        "token_authorization_rules",
-        token_authorization_rules::id(),
-        None,
-    )
+    ProgramTest::new("mpl_token_auth_rules", mpl_token_auth_rules::id(), None)
 }
